requirements.txt
flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.8.0